+++ lib/include/mb_mgr_job_api.h
@@ -2617,23 +2617,8 @@ SUBMIT_JOB_HASH_EX(IMB_MGR *state, IMB_JOB *job, const IMB_HASH_ALG hash_alg)
         MB_MGR_SHA_256_OOO *sha_256_ooo = state->sha_256_ooo;
         MB_MGR_SHA_512_OOO *sha_384_ooo = state->sha_384_ooo;
         MB_MGR_SHA_512_OOO *sha_512_ooo = state->sha_512_ooo;
-#ifdef SUBMIT_JOB_SHA3_224
-        MB_MGR_SHA3_OOO *sha3_224_ooo = state->sha3_224_ooo;
-#endif
-#ifdef SUBMIT_JOB_SHA3_256
-        MB_MGR_SHA3_OOO *sha3_256_ooo = state->sha3_256_ooo;
-#endif
-#ifdef SUBMIT_JOB_SHA3_384
-        MB_MGR_SHA3_OOO *sha3_384_ooo = state->sha3_384_ooo;
-#endif
-#ifdef SUBMIT_JOB_SHA3_512
-        MB_MGR_SHA3_OOO *sha3_512_ooo = state->sha3_512_ooo;
-#endif
-#ifdef SUBMIT_JOB_SHAKE128
-        MB_MGR_SHA3_OOO *shake128_ooo = state->shake128_ooo;
-#endif
-#ifdef SUBMIT_JOB_SHAKE256
-        MB_MGR_SHA3_OOO *shake256_ooo = state->shake256_ooo;
+#ifdef SUBMIT_JOB_SHA3
+        MB_MGR_SHA3_OOO *sha3_ooo = state->sha3_ooo;
 #endif
 #if (defined(SAFE_LOOKUP) || defined(AVX512))
         MB_MGR_SNOW3G_OOO *snow3g_uia2_ooo = state->snow3g_uia2_ooo;
@@ -2770,40 +2755,16 @@ SUBMIT_JOB_HASH_EX(IMB_MGR *state, IMB_JOB *job, const IMB_HASH_ALG hash_alg)
         case IMB_AUTH_SM3:
                 return SUBMIT_JOB_SM3(job);
         case IMB_AUTH_SHA3_224:
-#ifdef SUBMIT_JOB_SHA3_224
-                return SUBMIT_JOB_SHA3_224(sha3_224_ooo, job);
-#else
-                return submit_job_sha3(state, job, IMB_AUTH_SHA3_224);
-#endif
         case IMB_AUTH_SHA3_256:
-#ifdef SUBMIT_JOB_SHA3_256
-                return SUBMIT_JOB_SHA3_256(sha3_256_ooo, job);
-#else
-                return submit_job_sha3(state, job, IMB_AUTH_SHA3_256);
-#endif
         case IMB_AUTH_SHA3_384:
-#ifdef SUBMIT_JOB_SHA3_384
-                return SUBMIT_JOB_SHA3_384(sha3_384_ooo, job);
-#else
-                return submit_job_sha3(state, job, IMB_AUTH_SHA3_384);
-#endif
         case IMB_AUTH_SHA3_512:
-#ifdef SUBMIT_JOB_SHA3_512
-                return SUBMIT_JOB_SHA3_512(sha3_512_ooo, job);
-#else
-                return submit_job_sha3(state, job, IMB_AUTH_SHA3_512);
-#endif
         case IMB_AUTH_SHAKE128:
-#ifdef SUBMIT_JOB_SHAKE128
-                return SUBMIT_JOB_SHAKE128(shake128_ooo, job);
-#else
-                return submit_job_sha3(state, job, IMB_AUTH_SHAKE128);
-#endif
         case IMB_AUTH_SHAKE256:
-#ifdef SUBMIT_JOB_SHAKE256
-                return SUBMIT_JOB_SHAKE256(shake256_ooo, job);
+#ifdef SUBMIT_JOB_SHA3
+                /* the sponge rate and padding follow from job->hash_alg */
+                return SUBMIT_JOB_SHA3(sha3_ooo, job);
 #else
-                return submit_job_sha3(state, job, IMB_AUTH_SHAKE256);
+                return submit_job_sha3(state, job, job->hash_alg);
 #endif
         case IMB_AUTH_HMAC_SHA3_224:
 #ifdef SUBMIT_JOB_HMAC_SHA3_224
@@ -2867,23 +2828,8 @@ FLUSH_JOB_HASH_EX(IMB_MGR *state, IMB_JOB *job, const IMB_HASH_ALG hash_alg)
         MB_MGR_SHA_256_OOO *sha_256_ooo = state->sha_256_ooo;
         MB_MGR_SHA_512_OOO *sha_384_ooo = state->sha_384_ooo;
         MB_MGR_SHA_512_OOO *sha_512_ooo = state->sha_512_ooo;
-#ifdef FLUSH_JOB_SHA3_224
-        MB_MGR_SHA3_OOO *sha3_224_ooo = state->sha3_224_ooo;
-#endif
-#ifdef FLUSH_JOB_SHA3_256
-        MB_MGR_SHA3_OOO *sha3_256_ooo = state->sha3_256_ooo;
-#endif
-#ifdef FLUSH_JOB_SHA3_384
-        MB_MGR_SHA3_OOO *sha3_384_ooo = state->sha3_384_ooo;
-#endif
-#ifdef FLUSH_JOB_SHA3_512
-        MB_MGR_SHA3_OOO *sha3_512_ooo = state->sha3_512_ooo;
-#endif
-#ifdef FLUSH_JOB_SHAKE128
-        MB_MGR_SHA3_OOO *shake128_ooo = state->shake128_ooo;
-#endif
-#ifdef FLUSH_JOB_SHAKE256
-        MB_MGR_SHA3_OOO *shake256_ooo = state->shake256_ooo;
+#ifdef FLUSH_JOB_SHA3
+        MB_MGR_SHA3_OOO *sha3_ooo = state->sha3_ooo;
 #endif
 #if (defined(SAFE_LOOKUP) || defined(AVX512))
         MB_MGR_SNOW3G_OOO *snow3g_uia2_ooo = state->snow3g_uia2_ooo;
@@ -2914,27 +2860,15 @@ FLUSH_JOB_HASH_EX(IMB_MGR *state, IMB_JOB *job, const IMB_HASH_ALG hash_alg)
         case IMB_AUTH_SHA_512:
                 return FLUSH_JOB_SHA512(sha_512_ooo, job);
         case IMB_AUTH_SHA3_224:
-#ifdef FLUSH_JOB_SHA3_224
-                return FLUSH_JOB_SHA3_224(sha3_224_ooo, job);
-#else
-                return NULL;
-#endif
         case IMB_AUTH_SHA3_256:
-#ifdef FLUSH_JOB_SHA3_256
-                return FLUSH_JOB_SHA3_256(sha3_256_ooo, job);
-#else
-                return NULL;
-#endif
         case IMB_AUTH_SHA3_384:
-#ifdef FLUSH_JOB_SHA3_384
-                return FLUSH_JOB_SHA3_384(sha3_384_ooo, job);
-#else
-                return NULL;
-#endif
         case IMB_AUTH_SHA3_512:
-#ifdef FLUSH_JOB_SHA3_512
-                return FLUSH_JOB_SHA3_512(sha3_512_ooo, job);
+        case IMB_AUTH_SHAKE128:
+        case IMB_AUTH_SHAKE256:
+#ifdef FLUSH_JOB_SHA3
+                return FLUSH_JOB_SHA3(sha3_ooo, job);
 #else
+                /* submitted through the synchronous path, nothing to flush */
                 return NULL;
 #endif
         case IMB_AUTH_AES_XCBC:
@@ -2962,14 +2896,6 @@ FLUSH_JOB_HASH_EX(IMB_MGR *state, IMB_JOB *job, const IMB_HASH_ALG hash_alg)
 #ifdef FLUSH_JOB_SNOW5G_NIA4_X2
         case IMB_AUTH_SNOW5G_NIA4:
                 return FLUSH_JOB_SNOW5G_NIA4_X2(snow5g_nia4_ooo);
-#endif
-#ifdef FLUSH_JOB_SHAKE128
-        case IMB_AUTH_SHAKE128:
-                return FLUSH_JOB_SHAKE128(shake128_ooo, job);
-#endif
-#ifdef FLUSH_JOB_SHAKE256
-        case IMB_AUTH_SHAKE256:
-                return FLUSH_JOB_SHAKE256(shake256_ooo, job);
 #endif
         case IMB_AUTH_HMAC_SHA3_224:
         case IMB_AUTH_HMAC_SHA3_256: