+++ lib/include/ooo_mgr_reset.h
@@ -69,14 +69,6 @@ ooo_mgr_sha3_reset(void *p_ooo_mgr, const unsigned num_lanes);
 
 IMB_DLL_LOCAL
 void
-ooo_mgr_shake128_reset(void *p_ooo_mgr, const unsigned num_lanes);
-
-IMB_DLL_LOCAL
-void
-ooo_mgr_shake256_reset(void *p_ooo_mgr, const unsigned num_lanes);
-
-IMB_DLL_LOCAL
-void
 ooo_mgr_des_reset(void *p_ooo_mgr, const unsigned num_lanes);
 
 IMB_DLL_LOCAL